requirements.txt
newspaper3k
flask
gunicorn
dnspython

